storage/mysql: Refactor to move mysql decoding logic to mz-mysql-util
