+++ src/mysql-util/Cargo.toml
@@ -11,6 +11,7 @@ workspace = true
 
 [dependencies]
 anyhow = "1.0.66"
+chrono = { version = "0.4.23", default-features = false, features = ["std"] }
 indexmap = { version = "1.9.1", default-features = false, features = ["std"] }
 itertools = "0.10.5"
 mz-cloud-resources = { path = "../cloud-resources" }
@@ -18,6 +19,7 @@ mz-ore = { path = "../ore", features = ["async"] }
 mz-proto = { path = "../proto" }
 mz-repr = { path = "../repr" }
 mz-ssh-util = { path = "../ssh-util" }
+mysql_common = { version = "0.31.0", default-features = false, features = ["chrono"] }
 mysql_async = { version = "0.33.0", default-features = false, features = ["minimal", "tracing"] }
 once_cell = "1.16.0"
 prost = { version = "0.11.3", features = ["no-recursion-limit"] }
